requirements.txt
discord.py==2.3.2
psycopg2-binary==2.9.9
sqlalchemy==1.4.23
python-dotenv==0.19.0
openai==1.12.0
pinecone-client==3.0.2
sentence-transformers==2.5.1
fastapi==0.68.1
uvicorn==0.15.0
pydantic>=1.9.0,<3.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
aiohttp==3.8.1 